pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]


[project]
name = "my-project"
version = "0.1.0"
description = "Type your project description here."
readme = "README.md"
requires-python = ">=3.13"
authors = [{ name = "Your Name" }]

dependencies = [
    "omegaconf>=2.3.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.1.0",
]

[dependency-groups]
dev = [
    "commitizen>=4.13.9",
    "interrogate>=1.7.0",
    "pip-audit>=2.10.0",
    "pre-commit>=4.5.1",
    "pyright>=1.1.408",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "python-semantic-release>=10.5.3",
    "ruff>=0.15.2",
]

# =============================================================================
# Ruff — Linting & Formatting
# =============================================================================

[tool.ruff]
target-version = "py313"
line-length = 100
src = ["src"]
fix = true

[tool.ruff.lint]
select = [
    # --- Core ---
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "N",      # pep8-naming
    "UP",     # pyupgrade
    # --- Bug / Code-smell ---
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "SIM",    # flake8-simplify
    "PL",     # pylint
    "RUF",    # ruff-specific
    "PIE",    # flake8-pie
    "PERF",   # perflint
    "RET",    # flake8-return
    # --- Documentation / Typing ---
    "D",      # pydocstyle
    "ANN",    # flake8-annotations
    "TC",     # flake8-type-checking
    # --- Security ---
    "S",      # flake8-bandit (replaces standalone bandit)
    # --- Style ---
    "ARG",    # flake8-unused-arguments
    "A",      # flake8-builtins (prevent shadowing builtins)
    "ISC",    # flake8-implicit-str-concat
    "T20",    # flake8-print (catch stray print() calls)
    "PTH",    # flake8-use-pathlib
    "DTZ",    # flake8-datetimez
    "RSE",    # flake8-raise
    "TRY",    # tryceratops (exception handling)
    "LOG",    # flake8-logging
    # --- Testing ---
    "PT",     # flake8-pytest-style
]
ignore = [
    "E501",   # line length — handled by ruff formatter
    "D100",   # module docstrings — interrogate covers doc coverage
    "D104",   # __init__.py docstrings — unnecessary boilerplate
    "TRY003", # long exception messages — too strict for real-world code
    "ISC001", # implicit string concat single-line — conflicts with ruff formatter
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = [
    "D",       # no docstring requirements in tests
    "ANN",     # no type-annotation requirements in tests
    "S101",    # allow assert statements
    "ARG",     # allow unused args (pytest fixtures)
    "PLR2004", # allow magic numbers in assertions
    "PT011",   # allow broad pytest.raises()
]

[tool.ruff.lint.pylint]
max-args = 8

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true


# =============================================================================
# Interrogate — Docstring Coverage
# =============================================================================

[tool.interrogate]
fail-under = 70
ignore-init-method = true
ignore-init-module = true
ignore-module = true
exclude = ["tests"]


# =============================================================================
# Pyright — Static Type Checking
# =============================================================================

[tool.pyright]
typeCheckingMode = "strict"
venvPath = "."
venv = ".venv"
include = ["src"]
reportMissingTypeStubs = false


# =============================================================================
# Pytest
# =============================================================================

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "-ra",
]
filterwarnings = ["error"]


# =============================================================================
# Coverage
# =============================================================================

[tool.coverage.run]
source = ["src"]
branch = true

[tool.coverage.report]
fail_under = 80
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    'if __name__ == "__main__"',
]


# =============================================================================
# Semantic Release — CI version bumps, changelog, releases
# =============================================================================

[tool.semantic_release]
version_toml = ["pyproject.toml:project.version"]
build_command = "uv build"
commit_parser = "conventional"
commit_message = "chore(release): v{version}"
commit_author = "semantic-release <semantic-release@ci>"
tag_format = "v{version}"

[tool.semantic_release.branches.main]
match = "(main|master)"
prerelease = false

[tool.semantic_release.branches.alpha]
match = ".*"
prerelease_token = "alpha"
prerelease = true


# =============================================================================
# Commitizen — Local commit message format enforcement only
# =============================================================================

[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "v$version"
